requirements.txt
PyJWT==2.4.0
Flask==2.1.2
Werkzeug==2.1.2